Breaking apart NodeTranslatable into Entity and Content paradigm implementations. Adding EntityFieldTranslatableBase to handle future entity types.
