+++ src/Drupal/Translatable/Entity/NodeTranslatable.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * @file
+ * Defines a Node translatable compatible with Entity Translation module.
+ */
+
+namespace EntityXliff\Drupal\Translatable\Entity;
+
+use EntityXliff\Drupal\Translatable\EntityFieldTranslatableBase;
+
+
+/**
+ * Class NodeTranslatable
+ * @package EntityXliff\Drupal\Translatable\Entity
+ */
+class NodeTranslatable extends EntityFieldTranslatableBase {
+
+}
